Role/Region controller tests are actually working for real now
The last version wouldn't pass from a clean database. It would fail the first
time it was run if the database was clean - and then would pass every time after
that. That wasn't good enough.

There are lots of users and a role left over after the tests run; we'll have to
find out where the leak is.