Added a Codio button for compiling assets.
It cleans out the public/assets/ directory and recompiles assets.
This makes it easier to profile load times, as it's more production-like,
but it means that you have to recompile changes to CSS or JavaScript.

Recompiling assets is faster for load times, but not necessarily appropriate
for a development environment unless you're doing some profiling. I'll have
to turn off precompiling when I'm done.